introduce post caching
Reading and formatting every post in a thread each time it is accessed
is troublesome for threads with many posts, so add a 'postcache' file
which is updated when a new post is made.

Also ensure new threads and this postcache file have -group- write
permissions when they are created.